squash commits, move enum into resize.c
